feat(#181): split header contents
