+++ apps/frontend/src/components/Header/LandingHeader.tsx
@@ -1,0 +1,14 @@
+import { Button } from '../Button';
+
+export const LandingHeader = () => {
+  return (
+    <div className="flex items-center gap-2">
+      <Button round to="/signup">
+        Get Started
+      </Button>
+      <Button variant={'secondary'} round to="/signin">
+        Sign In
+      </Button>
+    </div>
+  );
+};
