+++ apps/frontend/src/components/Header/contents/index.ts
@@ -1,0 +1,6 @@
+export * from './ImportContents';
+export * from './ExportContents';
+export * from './ShareContents';
+export * from './VersionContents';
+export * from './SettingsContents';
+export * from './NotificationContents';
